requirements.txt
pre-commit==3.7.1
eth-ape~=0.7.0
ape-solidity==0.7.3
